Upgrade to Angular 19 and modernize various dependencies. (#7)
* Upgrade to Angular 19 and modernize various dependencies.

Updated Angular packages, TypeScript, and other dependencies to the latest compatible versions. Refactored code for compatibility, simplified provider return syntax, and added a "clear cache" feature to the demo app. Adjusted build tools to use the updated Angular build system.

* Refactor caching mechanism to support no-cache option

Introduced `NoCacheService` as an alternative to `RebrickableCacheService` with an abstract base class. Updated the `provideRebrickable` function to use `NoCacheService` by default and replaced local storage caching with IndexedDB in the demo setup.

* Refactor caching mechanism to support no-cache option

Introduced `NoCacheService` as an alternative to `RebrickableCacheService` with an abstract base class. Updated the `provideRebrickable` function to use `NoCacheService` by default and replaced local storage caching with IndexedDB in the demo setup.